RENAME (Chapter4, 5) : main클래스의 이름 수정
