+++ Chapter5/src/main/java/com/nhnacademy/Jminsoo/excercise7/Excercise5_7.java
@@ -1,8 +1,8 @@
-package com.nhnacademy.jminsoo.excercise8;
+package com.nhnacademy.Jminsoo.excercise7;
 
-public class Main {
+public class Excercise5_7 {
     public static void main(String[] args) {
-        AddtionQuiz addtionQuiz = new AddtionQuiz(10);
+        Quiz addtionQuiz = new Quiz(10);
 
         addtionQuiz.quizGenerate();
         addtionQuiz.quizControl();
